+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Features:
   - Import: Mark import as debug -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/47)
   - Milestone View: Add functionality that displays the latest results of the tests from the test runs -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/11)
   - List of predefined Resolutions -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/26)
+  - Add posibility to Finish Test Run Manually -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/21)
+
+Bugfixes:
+  - Cannot remove milestone from TestRun -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/50)
 
 ## 0.3.4 (2019-12-10)